Holy hell, that was a programming gauntlet.  So there's a quirk with the javascript table control that requires table head structure to be in place before it can be auto populated.
This can be problematic for dynamically making the head.  I'll need to write a script that loops through head data and builds a template for the datatable control to populate.

Also, I had a level of success in discovering that I don't need to parse the data like a Nazi, I can throw it into a string array and send it to the client to be sorted.  This technique probably works well with data dumps, not so much when you have parameters for your data.